feat: New value type option class

+++ dhis-2/dhis-api/src/test/java/org/hisp/dhis/common/ValueTypeOptionsTest.java
@@ -46,7 +46,7 @@ public class ValueTypeOptionsTest
         ftvo.setMaxFileSize( 100L );
         ftvo.setAllowedContentTypes( ImmutableSet.of( "jpg" ) );
 
-        assertEquals( 100L, ftvo.getMaxFileSize().longValue() );
+        assertEquals( 100L, ftvo.getMaxFileSize() );
         assertTrue( ftvo.getAllowedContentTypes().contains( "jpg" ) );
     }
 }
